Partial implementation of inproc ngc compiler and some more minor refactoring
Beginnings of implementation for inprocess compiler so that we can
prerender routes just by being given a tsconfig.json. Ultimately this is
to support projects that use the Angular CLI and cannot add custom
webpack configurations to add new targets for SSR. This way, we can
simply add a new build step which invokes an angular-ssr tool, accepts a
tsconfig.json file and spits out all the prerendered routes.

This also flattens out the directory structure of the project a bit, and
places all the custom exception types in a single file, exception.ts,
instead of spreading them throughout the source tree.