Refactor our relationship with global window and document objects, disable jest-environment-jsdom
Add a zone-based map for global window and document objects. This means
that each time we instantiate a DocumentContainer (which hosts the real
DOM implementation), we map the current zone.js zone to that container,
so that subsequent requests for the global window or document objects
will map to the correct DOM instance in the running application.

As part of this, disable jsdom. I didn't realize this was a default when
running jest. Before disabling it, it used jsdom to define global
window and document objects which caused bizarre behaviour.