Major overhaul of ApplicationBuilder API to separate out Application from ApplicationBuilder
Previously, ApplicationBuilder was called builder but had no build()
method. This didn't make a lot of sense, so now all ApplicationBuilder
implementations have a build() function that returns an Application,
which itself can be used to discover routes, pre-render, and demand
render URLs. The Application object returned from build() must be
disposed when you are finished with it using application.dispose().